feat: add Prime method to prime request cache
Implementation is copied from the existing dataloaden library, see link
below. The passed value needs to be copied before inserted in the cache
to prevent common developer errors (passing a pointer from a loop
variable), so a new config option is added.

Instead of silently failing if Copy is not specified, and to not make a
breaking change, the new Prime method panics if this method is not
specified.

https://github.com/vektah/dataloaden/blob/557c25d5f7dcd7d6fc6aa65ea0781e214b9784b9/example/userloader_gen.go#LL147-L161C2